+++ .github/workflows/qc-and-test.yaml
@@ -17,9 +17,9 @@ jobs:
 
       - name: Unit tests [pytest]
         run: |
-          pytest --cov=jupyter_utils --cov-report term-missing
+          pytest --cov=magicbox --cov-report term-missing
           coverage report --format=markdown -m >> $GITHUB_STEP_SUMMARY
 
       - name: QC [ruff]
         if: ${{!cancelled()}}
-        run: ruff check jupyter_utils
+        run: ruff check magicbox
