--- jupyter_utils/__init__.py
@@ -1,3 +1,0 @@
-"""Package for jupyter_utils."""
-
-__version__ = "0.3.2"
