+++ magicbox/__init__.py
@@ -1,0 +1,3 @@
+"""Package for magicbox."""
+
+__version__ = "0.3.3"
