+++ magicbox/exceptions.py
