+++ magicbox/functions.py
