--- requirements.in
@@ -1,2 +1,0 @@
-notebook
-ipykernel
