+++ requirements.txt
@@ -4,280 +4,3 @@
 #
 #    pip-compile --output-file=requirements.txt requirements.in
 #
-anyio==4.3.0
-    # via
-    #   httpx
-    #   jupyter-server
-argon2-cffi==23.1.0
-    # via jupyter-server
-argon2-cffi-bindings==21.2.0
-    # via argon2-cffi
-arrow==1.3.0
-    # via isoduration
-asttokens==2.4.1
-    # via stack-data
-async-lru==2.0.4
-    # via jupyterlab
-attrs==23.2.0
-    # via
-    #   jsonschema
-    #   referencing
-babel==2.15.0
-    # via jupyterlab-server
-beautifulsoup4==4.12.3
-    # via nbconvert
-bleach==6.1.0
-    # via nbconvert
-certifi==2024.2.2
-    # via
-    #   httpcore
-    #   httpx
-    #   requests
-cffi==1.16.0
-    # via argon2-cffi-bindings
-charset-normalizer==3.3.2
-    # via requests
-comm==0.2.2
-    # via ipykernel
-debugpy==1.8.1
-    # via ipykernel
-decorator==5.1.1
-    # via ipython
-defusedxml==0.7.1
-    # via nbconvert
-executing==2.0.1
-    # via stack-data
-fastjsonschema==2.19.1
-    # via nbformat
-fqdn==1.5.1
-    # via jsonschema
-h11==0.14.0
-    # via httpcore
-httpcore==1.0.5
-    # via httpx
-httpx==0.27.0
-    # via jupyterlab
-idna==3.7
-    # via
-    #   anyio
-    #   httpx
-    #   jsonschema
-    #   requests
-ipykernel==6.29.4
-    # via
-    #   -r requirements.in
-    #   jupyterlab
-ipython==8.24.0
-    # via ipykernel
-isoduration==20.11.0
-    # via jsonschema
-jedi==0.19.1
-    # via ipython
-jinja2==3.1.4
-    # via
-    #   jupyter-server
-    #   jupyterlab
-    #   jupyterlab-server
-    #   nbconvert
-json5==0.9.25
-    # via jupyterlab-server
-jsonpointer==2.4
-    # via jsonschema
-jsonschema[format-nongpl]==4.22.0
-    # via
-    #   jupyter-events
-    #   jupyterlab-server
-    #   nbformat
-jsonschema-specifications==2023.12.1
-    # via jsonschema
-jupyter-client==8.6.1
-    # via
-    #   ipykernel
-    #   jupyter-server
-    #   nbclient
-jupyter-core==5.7.2
-    # via
-    #   ipykernel
-    #   jupyter-client
-    #   jupyter-server
-    #   jupyterlab
-    #   nbclient
-    #   nbconvert
-    #   nbformat
-jupyter-events==0.10.0
-    # via jupyter-server
-jupyter-lsp==2.2.5
-    # via jupyterlab
-jupyter-server==2.14.0
-    # via
-    #   jupyter-lsp
-    #   jupyterlab
-    #   jupyterlab-server
-    #   notebook
-    #   notebook-shim
-jupyter-server-terminals==0.5.3
-    # via jupyter-server
-jupyterlab==4.1.8
-    # via notebook
-jupyterlab-pygments==0.3.0
-    # via nbconvert
-jupyterlab-server==2.27.1
-    # via
-    #   jupyterlab
-    #   notebook
-markupsafe==2.1.5
-    # via
-    #   jinja2
-    #   nbconvert
-matplotlib-inline==0.1.7
-    # via
-    #   ipykernel
-    #   ipython
-mistune==3.0.2
-    # via nbconvert
-nbclient==0.10.0
-    # via nbconvert
-nbconvert==7.16.4
-    # via jupyter-server
-nbformat==5.10.4
-    # via
-    #   jupyter-server
-    #   nbclient
-    #   nbconvert
-nest-asyncio==1.6.0
-    # via ipykernel
-notebook==7.1.3
-    # via -r requirements.in
-notebook-shim==0.2.4
-    # via
-    #   jupyterlab
-    #   notebook
-overrides==7.7.0
-    # via jupyter-server
-packaging==24.0
-    # via
-    #   ipykernel
-    #   jupyter-server
-    #   jupyterlab
-    #   jupyterlab-server
-    #   nbconvert
-pandocfilters==1.5.1
-    # via nbconvert
-parso==0.8.4
-    # via jedi
-pexpect==4.9.0
-    # via ipython
-platformdirs==4.2.1
-    # via jupyter-core
-prometheus-client==0.20.0
-    # via jupyter-server
-prompt-toolkit==3.0.43
-    # via ipython
-psutil==5.9.8
-    # via ipykernel
-ptyprocess==0.7.0
-    # via
-    #   pexpect
-    #   terminado
-pure-eval==0.2.2
-    # via stack-data
-pycparser==2.22
-    # via cffi
-pygments==2.18.0
-    # via
-    #   ipython
-    #   nbconvert
-python-dateutil==2.9.0.post0
-    # via
-    #   arrow
-    #   jupyter-client
-python-json-logger==2.0.7
-    # via jupyter-events
-pyyaml==6.0.1
-    # via jupyter-events
-pyzmq==26.0.3
-    # via
-    #   ipykernel
-    #   jupyter-client
-    #   jupyter-server
-referencing==0.35.1
-    # via
-    #   jsonschema
-    #   jsonschema-specifications
-    #   jupyter-events
-requests==2.31.0
-    # via jupyterlab-server
-rfc3339-validator==0.1.4
-    # via
-    #   jsonschema
-    #   jupyter-events
-rfc3986-validator==0.1.1
-    # via
-    #   jsonschema
-    #   jupyter-events
-rpds-py==0.18.0
-    # via
-    #   jsonschema
-    #   referencing
-send2trash==1.8.3
-    # via jupyter-server
-six==1.16.0
-    # via
-    #   asttokens
-    #   bleach
-    #   python-dateutil
-    #   rfc3339-validator
-sniffio==1.3.1
-    # via
-    #   anyio
-    #   httpx
-soupsieve==2.5
-    # via beautifulsoup4
-stack-data==0.6.3
-    # via ipython
-terminado==0.18.1
-    # via
-    #   jupyter-server
-    #   jupyter-server-terminals
-tinycss2==1.3.0
-    # via nbconvert
-tornado==6.4
-    # via
-    #   ipykernel
-    #   jupyter-client
-    #   jupyter-server
-    #   jupyterlab
-    #   notebook
-    #   terminado
-traitlets==5.14.3
-    # via
-    #   comm
-    #   ipykernel
-    #   ipython
-    #   jupyter-client
-    #   jupyter-core
-    #   jupyter-events
-    #   jupyter-server
-    #   jupyterlab
-    #   matplotlib-inline
-    #   nbclient
-    #   nbconvert
-    #   nbformat
-types-python-dateutil==2.9.0.20240316
-    # via arrow
-typing-extensions==4.11.0
-    # via ipython
-uri-template==1.3.0
-    # via jsonschema
-urllib3==2.2.1
-    # via requests
-wcwidth==0.2.13
-    # via prompt-toolkit
-webcolors==1.13
-    # via jsonschema
-webencodings==0.5.1
-    # via
-    #   bleach
-    #   tinycss2
-websocket-client==1.8.0
-    # via jupyter-server
